application-generator/src/content/python/modules/contact_us/requirements.txt
alembic==1.7.6
aniso8601==9.0.1
apifairy==0.9.1
apispec==5.2.2
blinker==1.4
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Cors==3.0.10
Flask-DotEnv==0.1.2
Flask-HTTPAuth==4.7.0
Flask-Mail==0.9.1
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Moment==1.0.5
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-sqlalchemy==0.28.0
mysql-connector-python==8.0.29
packaging==21.3
protobuf==4.21.8
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
webargs==8.2.0
Werkzeug==2.0.3